Improved typing of database stuff

+++ src/dao/sqls/kara.ts
@@ -58,13 +58,13 @@ ${offsetClause}
 export const getYears = 'SELECT year, karacount::integer FROM all_years ORDER BY year';
 
 export const selectBaseStats = `SELECT
-(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[2]) AS singers,
-(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[8]) AS songwriters,
-(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[4]) AS creators,
-(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[6]) AS authors,
-(SELECT COUNT(pk_kid) FROM kara) AS karas,
-(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[5] AND karacount::text NOT LIKE '{"count": 0}' ) AS languages,
-(SELECT COUNT(pk_sid) FROM serie) AS series,
-(SELECT SUM(mediasize) FROM kara) AS mediasize,
-(SELECT SUM(duration) FROM kara) AS duration;
+(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[2])::integer AS singers,
+(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[8])::integer AS songwriters,
+(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[4])::integer AS creators,
+(SELECT COUNT(tid) FROM all_tags WHERE types @> ARRAY[6])::integer AS authors,
+(SELECT COUNT(pk_kid) FROM kara)::integer AS karas,
+(SELECT COUNT(tid)::integer FROM all_tags WHERE types @> ARRAY[5] AND karacount::text NOT LIKE '{"count": 0}' ) AS languages,
+(SELECT COUNT(pk_sid) FROM serie)::integer AS series,
+(SELECT SUM(mediasize) FROM kara)::integer AS mediasize,
+(SELECT SUM(duration) FROM kara)::integer AS duration;
 `;
